add sql scripts

+++ src/main/java/ru/jimaltair/currencyconverter/repository/CurrencyRateRepository.java
@@ -3,5 +3,5 @@ package ru.jimaltair.currencyconverter.repository;
 import org.springframework.data.repository.CrudRepository;
 import ru.jimaltair.currencyconverter.entity.CurrencyRate;
 
-public interface CurrencyRateRepository extends CrudRepository<CurrencyRate, Integer> {
+public interface CurrencyRateRepository extends CrudRepository<CurrencyRate, Long> {
 }
